third_party/WebKit/LayoutTests/fast/forms/textarea/textarea-select.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<textarea id="text"></textarea>
<script>
description("This test calls textarea.select and confirms selectionStart/End.");

text.value = "hello\n";
text.select();
shouldBe("text.selectionStart", "0");
shouldBe("text.selectionEnd", "6");
</script>
</body>
</html>
