third_party/blink/web_tests/fast/forms/textarea/textarea-select-expected.txt

This test calls textarea.select and confirms selectionStart/End.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS text.selectionStart is 0
PASS text.selectionEnd is 6
PASS successfullyParsed is true

TEST COMPLETE
